Move a lot of files around.
Tests will fail because none of the paths have been updated.
